fixed routing for buy and sell info page
