+++ lib/main.dart
@@ -277,13 +277,14 @@ class MyAppState extends State<MyApp> with WidgetsBindingObserver {
                       isBody: true,
                       entityID: temp.split("/putentity/body/")[1],
                       cookie: widget.bloc.getSessionIdHeader()));
-            } else if (temp.startsWith("/buyandsell/info")) {
-              return _buildRoute(
-                  settings,
-                  BuyAndSellInfoPage(
-                      post: widget.bloc.buynSellPostBloc
-                          .getBuynSellPost(temp.split("/buyandsell/info")[1])));
-            } else {
+            } //else if (temp.startsWith("/buyandsell/info")) {
+              // return _buildRoute(
+              //     settings,
+              //     BuyAndSellInfoPage(
+              //         post: widget.bloc.buynSellPostBloc
+              //             .getBuynSellPost(temp.split("/buyandsell/info")[1])));
+          //  }
+          else {
               switch (settings.name) {
                 case "/":
                   return _buildRoute(
